+++ src/util/binary.js
